ml-model/deployment/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
pillow>=9.0.0
boto3>=1.26.0
